testing..

+++ examples/mlperf/model_train.py
@@ -1456,7 +1456,7 @@ def train_llama3(llama2_70b_lora:bool=False):
   for p in optim.params:
     grad_dtype = dtypes.bfloat16 if p.dtype == FP8_DTYPE else p.dtype
     # grads are sharded if using FSDP
-    p.grad = p.empty_like(dtype=grad_dtype).contiguous()
+    p.grad = p.empty_like().contiguous()
   grads = [p.grad for p in optim.params]
 
   scheduler = CosineAnnealingLRWithWarmup(optim, opt_base_learning_rate, opt_end_learning_rate, opt_learning_rate_warmup_steps, opt_learning_rate_decay_steps)
